// packages/app/api/generated/models/CalendarSlot.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CalendarSlot
 */
export interface CalendarSlot {
    /**
     * 
     * @type {string}
     * @memberof CalendarSlot
     */
    id?: string;
    /**
     * 
     * @type {Date}
     * @memberof CalendarSlot
     */
    lessonTime?: Date;
    /**
     * 
     * @type {boolean}
     * @memberof CalendarSlot
     */
    available?: boolean;
    /**
     * 
     * @type {string}
     * @memberof CalendarSlot
     */
    tutorId?: string;
}

/**
 * Check if a given object implements the CalendarSlot interface.
 */
export function instanceOfCalendarSlot(value: object): value is CalendarSlot {
    return true;
}

export function CalendarSlotFromJSON(json: any): CalendarSlot {
    return CalendarSlotFromJSONTyped(json, false);
}

export function CalendarSlotFromJSONTyped(json: any, ignoreDiscriminator: boolean): CalendarSlot {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'lessonTime': json['lessonTime'] == null ? undefined : (new Date(json['lessonTime'])),
        'available': json['available'] == null ? undefined : json['available'],
        'tutorId': json['tutorId'] == null ? undefined : json['tutorId'],
    };
}

export function CalendarSlotToJSON(json: any): CalendarSlot {
    return CalendarSlotToJSONTyped(json, false);
}

export function CalendarSlotToJSONTyped(value?: CalendarSlot | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'lessonTime': value['lessonTime'] == null ? value['lessonTime'] : value['lessonTime'].toISOString(),
        'available': value['available'],
        'tutorId': value['tutorId'],
    };
}